+++ src/plugin/host-process.cpp
@@ -234,8 +234,17 @@ bool GroupHost::running() {
     // process is still running, however Boost.Process does not allow you to do
     // the same thing for a process that's not a direct child if this process.
     // When using plugin groups we'll have to manually check whether the PID
-    // returned by the group host process is still active.
-    return kill(host_pid, 0) == 0;
+    // returned by the group host process is still active. We sadly can't use
+    // `kill()` for this as that provides no way to distinguish between active
+    // processes and zombies, and a terminated group host process will always be
+    // left as a zombie process. If the process is active, then
+    // `/proc/<pid>/{cwd,exe,root}` will be valid symlinks.
+    try {
+        fs::canonical("/proc/" + std::to_string(host_pid) + "/exe");
+        return true;
+    } catch (const fs::filesystem_error&) {
+        return false;
+    }
 }
 
 void GroupHost::terminate() {